perf(review): address PR #675 review
- Move the timing-dependent harness meta-tests (measureMs, canary,
  self-consistency) out of the parallel default unit suite into a serialized
  harness.perf.test.ts. In the default suite the `< GROWTH_BOUND` assertion on
  real timings could flake under CPU contention and block a merge -- the
  opposite of the informational-first design. The deterministic tests (schedule,
  recorder, gate) stay in the unit suite and still gate. Closures resized to a
  few ms (pure arithmetic loops, no large arrays) so single-sample timing is
  stable even serialized.
- Guard growthFactor's ratio against a zero-time sample (floor each timing at
  MIN_MS) so a too-fast closure degrades to a finite number instead of
  Infinity/NaN poisoning the median.
- perfEnforceEnabled(): treat PERF_ENFORCE=0/false/"" as off, so a value like
  "0" cannot accidentally turn gating on (plain Boolean("0") is true). Added a
  test for it.
- Work-happened assertion now checks the SMALL fixture too, guarding both
  measurement points.
- Actually ignore perf-results.jsonl: the entry has to live in the tracked root
  .gitignore -- a packages/agency-lang/.gitignore is itself swept by the root
  `.*` dotfile rule and can never be committed.
